requirements.txt
django
ShopifyAPI
djangorestframework
markdown
django-filter
django-rest-swagger
mysqlclient
channels
daphne
